requirements.txt
# RAGAIMODEL project dependencies
llama-cpp-python>=0.1.129
python-dateutil>=2.8.2
openai>=0.27.4
networkx
PyYAML
torch
numpy
scikit-learn
langchain==0.3.24
chromadb==1.0.7
sentence-transformers==4.1.0
langchain-chroma>=0.1.0
langchain-huggingface>=0.1.0
accelerate-1.6.0 
bitsandbytes-0.42.0 
psutil-7.0.0
sentencepiece-0.2.0
#HF login and token for "read"
hf_xet-1.1.0
langchain>=0.1.0

# Optional: to enable Metal (GPU) acceleration on Apple Silicon (M1/M2) Macs
# Run the following in your terminal before installing llama-cpp-python:
# CMAKE_ARGS='-DLLAMA_METAL=on' FORCE_CMAKE=1 pip install --upgrade --force-reinstall llama-cpp-python
